requirements.txt
flask
flask-cors
flask_sqlalchemy
flask_login
werkzeug
mysqlclient
tensorflow
keras
scikit-learn
pandas
numpy
python-dotenv
python-dateutil
reportlab
dateparser
pymysql
psycopg2-binary
gunicorn